+++ src/app/components/Hero/index.ts
