--- src/app/lib/hooks/index.ts
@@ -1,2 +1,0 @@
-export {useInnerWidth} from "./useInnerWidth"
-export {useLoader} from "./useLoader"
