+++ src/app/page.tsx
@@ -1,8 +1,6 @@
-import type { Metadata } from "next"
-
-import Hero from "./components/ui/Hero"
-import Featured from "./components/ui/Featured"
-import Gallery from "./components/ui/Gallery"
+import Hero from "./components/Hero"
+import Featured from "./components/Featured"
+import Gallery from "./components/Gallery"
 
 import cn from "clsx"
